test_test.py
import unittest

class TestApplicationLogic(unittest.TestCase):
    def test_add_register(self):
        # Create a mock DataFrame object
        class MockDataFrame:
            def __init__(self):
                self.registers = []

            def add_one(self, Register, reg):
                self.registers.append(reg)

            def commit(self):
                pass

        # Create an instance of the mock DataFrame
        df = MockDataFrame()

        # Call the application_logic function
        application_logic(df)

        # Assert that a new Register object was added to the DataFrame
        self.assertEqual(len(df.registers), 1)
        self.assertEqual(df.registers[0].user_agent, "example_user_agent")
        self.assertEqual(df.registers[0].fresh, True)

if __name__ == "__main__":
    unittest.main()